requirements.txt
tweepy
python-dotenv
speedtest-cli
plotly
